Move setup.py to ecog_setup.py to prevent issue with name conflict in colab
